FEAT: Do more refactoring
